capsicum: invoke Capsicum on FD/file conversion
Add calls to Capsicum to intercept FD/file conversion in the following
situations:
 - Any place where a file descriptor is converted to a struct file.
 - Any place where a new file descriptor that is derived from an
   existing file descriptor is installed into the fdtable.

For the first of these, Capsicum checks for a Capsicum capability
wrapper file and unwraps it to the underlying file, provided the
appropriate rights are available (otherwise, ERR_PTR(-ENOTCAPABLE)
is returned).

For the second, Capsicum checks whether the original file has
restricted rights associated with it.  If it does, it is replaced
with a Capsicum capability wrapper file before installation into
the fdtable.  This propagates the rights from the original FD to
the new FD, and particularly affects accept(2) and openat(2).

For openat(2) rights propagation in particular, the rights
associated with the dfd need to make their way through the code
in fs/namei.c to allow this.

The path walking code in fs/namei.c is also modified to enable
the O_BENEATH flag if the dfd is a Capsicum capability.

Signed-off-by: David Drysdale <drysdale@google.com>